Concurrency improvements and bug fixes and additional unit tests. Closes #14

+++ src/Strava/Model/AuthorizationResponse.cs
@@ -29,7 +29,7 @@ internal class AuthorizationResponse
 {
 #pragma warning disable IDE1006 // Naming Styles
     public string? token_type { get; init; }
-    public int? expires_at { get; init; }
+    public long? expires_at { get; init; }
     public int? expires_in { get; init; }
     public string? access_token { get; init; }
     public string? refresh_token { get; init; }
